feat:t-core: Inject AuthUserClient via interface fix:t-core, t-auth: Minor fixes for OpenAPI files
